fast-api-skeleton/content/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy-spanner